fix an issue, where an error gets thrown in an ajax call, when a pad is switched to non public and a password was never set

+++ src/Ethergroups/MainBundle/Controller/DefaultController.php
@@ -535,7 +535,7 @@ class DefaultController extends Controller {
 	    $em = $this->getDoctrine()->getManager();
 	    $repository = $em->getRepository('EthergroupsMainBundle:Pads');
 	    $pad = $repository->findOneBy(array('padid'=>$padid));
-	    $em->remove($pad);
+        if(isset($pad)) $em->remove($pad);
 	    $em->flush();
 	}
